Move existing config app tests to test folder
