+++ config/tests/test_schema.py
@@ -2,27 +2,9 @@ from unittest.mock import patch
 
 import pytest
 from django.http import Http404
-from django.test import TestCase
 
 from config.views.schema import schema
 from judgments.tests.fixtures import TestCaseWithMockAPI
-
-
-class TestCacheHeaders(TestCase):
-    def test_static_headers(self):
-        url = "/static/images/tna_logo.svg"
-        response = self.client.get(url)
-        assert response.headers["Cache-Control"] == "max-age=900, public"
-
-    def test_view_headers(self):
-        url = "/about-this-service"
-        response = self.client.get(url)
-        assert response.headers["Cache-Control"] == "max-age=900, public"
-
-    def test_no_cache_transactional_steps(self):
-        url = "/re-use-find-case-law-records/steps/organization"
-        response = self.client.get(url)
-        assert response.headers["Cache-Control"] == "max-age=0, no-cache, no-store, must-revalidate, public"
 
 
 class TestSchemas(TestCaseWithMockAPI):
